requirements.txt
asgiref==3.5.2
Django==4.0.6
pip==21.3.1
psycopg2==2.9.3
setuptools==60.2.0
sqlparse==0.4.2
tzdata==2022.1
wheel==0.37.1
flake8==5.0.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
docker==5.0.3
